+++ src/Webform.unit.js
@@ -78,6 +78,7 @@ import formWithDeeplyNestedConditionalComps from '../test/forms/formWithDeeplyNe
 import formWithValidation from '../test/forms/formWithValidation';
 import formWithNotAllowedTags from '../test/forms/formWithNotAllowedTags';
 import formWithValidateWhenHidden from '../test/forms/formWithValidateWhenHidden';
+import formWithEditGrid from '../test/forms/formWithEditGrid';
 import formWithSelectRadioUrlDataSource from '../test/forms/selectRadioUrlDataSource';
 const SpySanitize = sinon.spy(FormioUtils, 'sanitize');
 
@@ -372,6 +373,20 @@ describe('Webform tests', function() {
           assert.deepEqual(dataGrid.dataValue, dataGridData);
           done();
       }, 200);
+    }).catch((err) => done(err));
+  });
+
+  it('Should trigger validation for each row of data and edit grid', function(done) {
+    const formElement = document.createElement('div');
+    Formio.createForm(formElement, formWithEditGrid).then((form) => {
+      form.setSubmission({ data: 
+        { dataGrid: [{ textField: '' }, { textField: '' }],
+         editGrid: [{ number: '' }, { number: '' }]} 
+        });
+      setTimeout(() => {
+        assert.equal(form.errors.length, 4);
+        done();
+      }, 500);
     }).catch((err) => done(err));
   });