+++ src/components/number/fixtures/comp10.js
@@ -1,0 +1,19 @@
+export default {
+  components: [
+    {
+      "label": "Number",
+      "applyMaskOn": "change",
+      "mask": false,
+      "tableView": false,
+      "delimiter": false,
+      "requireDecimal": false,
+      "inputFormat": "plain",
+      "truncateMultipleSpaces": false,
+      "validateWhenHidden": false,
+      "key": "number",
+      "type": "number",
+      "input": true,
+      "defaultValue": "123.23"
+    }
+  ]
+}