+++ src/components/selectboxes/SelectBoxes.unit.js
@@ -13,6 +13,7 @@ import {
   comp7
 } from './fixtures';
 import wizardWithSelectBoxes from '../../../test/forms/wizardWithSelectBoxes';
+import {comp12} from "../radio/fixtures";
 
 describe('SelectBoxes Component', () => {
   it('Should build a SelectBoxes component', () => {
@@ -372,6 +373,78 @@ describe('SelectBoxes Component', () => {
     _.set(selectboxes, 'root._form.config.myApiUrl', 'test123');
     selectboxes.loadItems('{{form.config.myApiUrl}}', {}, {}, {});
     Formio.makeRequest = originalMakeRequest;
+  });
+
+  it('Should have correct submission data when setting the value property', (done) => {
+    const originalMakeRequest = Formio.makeRequest;
+    Formio.makeRequest = async ()=> {
+      return [
+        {
+          data: {
+            name: 'Bob',
+            referenceId: '1'
+          }
+        },
+        {
+          data: {
+            name: 'Tom',
+            referenceId: '2'
+          }
+        },
+        {
+          data: {
+            name: 'Joe',
+            referenceId: '3'
+          }
+        }
+      ]
+    }
+    const changeEvent = new Event('change');
+    Formio.createForm(document.createElement('div'), comp12, {}).then((form) => {
+      setTimeout(()=>{
+        const selectBoxesComponent = form.getComponent('selectBoxes');
+        selectBoxesComponent.refs.input[0].checked = true;
+        selectBoxesComponent.refs.input[0].dispatchEvent(changeEvent);
+        setTimeout(()=>{
+          assert.deepEqual(form.getValue().data, {selectBoxes : {Bob: true, Tom: false, Joe: false}});
+          done();
+        },200);
+        Formio.makeRequest = originalMakeRequest;
+      }, 200)
+    });
+  });
+
+  it('Should show validation errors when the value property is set', () => {
+    const originalMakeRequest = Formio.makeRequest;
+    Formio.makeRequest = async () => {
+      return [
+        {
+          data: {
+            name: 'Bob',
+            referenceId: '1'
+          }
+        },
+        {
+          data: {
+            name: 'Tom',
+            referenceId: '2'
+          }
+        },
+        {
+          data: {
+            name: 'Joe',
+            referenceId: '3'
+          }
+        }
+      ]
+    }
+    let newComp12 = _.cloneDeep(comp12);
+    _.set(newComp12.components[0], 'validate.required', true);
+    return Formio.createForm(document.createElement('div'), newComp12, {}).then((form) => {
+      const selectBoxesComponent = form.getComponent('selectBoxes');
+      assert.equal(selectBoxesComponent.errors.length, 1);
+      Formio.makeRequest = originalMakeRequest;
+    })
   });
 });